[#1707] Attach user profile to application logs
