--- lib/java/spring/events/src/main/java/co/airy/spring/events/config/EnableAsyncConfiguration.java
@@ -1,9 +1,0 @@
-package co.airy.spring.events.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.EnableAsync;
-
-@Configuration
-@EnableAsync
-public class EnableAsyncConfiguration {
-}
