+++ lib/java/spring/web/BUILD
@@ -10,9 +10,12 @@ custom_java_library(
         "//:jwt",
         "//:lombok",
         "//:spring",
+        "//backend/avro:http-log",
+        "//lib/java/kafka/schema:ops-application-logs",
         "//lib/java/log",
+        "//lib/java/spring/auth:spring-auth",
         "//lib/java/spring/core:spring-core",
-        "//lib/java/spring/events:spring-events",
+        "//lib/java/spring/kafka/core:spring-kafka-core",
         "@maven//:javax_servlet_javax_servlet_api",
         "@maven//:javax_xml_bind_jaxb_api",
         "@maven//:org_springframework_boot_spring_boot_autoconfigure",
